Enhance particle sensor reading and ISO evaluation functionality

+++ app/Database/Migrations/2025-08-29-063538_IsoLimits.php
@@ -16,7 +16,7 @@ class CreateIsoLimits extends Migration
                 'auto_increment' => true,
             ],
             'iso_class' => [
-                'type' => 'TINYINT',
+                'type' => 'INT',
                 'constraint' => 1,
                 'null' => false,
                 'unique' => true,
